Initial WebApp pushed to master-trainee-kpt
So, Basically This is a Shoppingcart in this website i used werkzeuf which is comprehensive WSGI web application library. I am using the folllowing technology to build these web application are HTML, CSS, Javascript. The working of website is like that these will going to be a only one page website throgh it i used to perform an operations like search items, add item into cart, cart is just in right hand side of items so if someone going to add item it will add in cart detail also but i have to make the javascript part, we can also remove the item and it should be reflect in cart-detail.